R - Endrer navn fra JournalforingMedBruker til Vurderbrukerinfo
