+++ src/test/kotlin/no/nav/eessi/pensjon/integrasjonstest/ConfigRestTemplateTest.kt
@@ -15,7 +15,7 @@ import no.nav.eessi.pensjon.eux.model.document.SedDokumentfiler
 import no.nav.eessi.pensjon.eux.model.document.SedVedlegg
 import no.nav.eessi.pensjon.gcp.GcpStorageService
 import no.nav.eessi.pensjon.integrasjonstest.saksflyt.JournalforingTestBase
-import no.nav.eessi.pensjon.journalforing.JournalforingMedBruker
+import no.nav.eessi.pensjon.journalforing.VurderBrukerInfo
 import no.nav.eessi.pensjon.journalforing.OpprettJournalpostRequest
 import no.nav.eessi.pensjon.journalforing.journalpost.JournalpostKlient
 import no.nav.eessi.pensjon.journalforing.saf.SafClient
@@ -66,7 +66,7 @@ internal class ConfigRestTemplateTest {
     private lateinit var sedSendtListener: SedSendtListener
 
     @MockkBean
-    private lateinit var journalforingutenbruker: JournalforingMedBruker
+    private lateinit var journalforingutenbruker: VurderBrukerInfo
 
     @MockkBean
     private lateinit var personService: PersonService
