+++ src/test/kotlin/no/nav/eessi/pensjon/integrasjonstest/SedSendtIntegrationTest.kt
@@ -7,7 +7,7 @@ import no.nav.eessi.pensjon.EessiPensjonJournalforingTestApplication
 import no.nav.eessi.pensjon.eux.klient.EuxKlientLib
 import no.nav.eessi.pensjon.eux.model.buc.Buc
 import no.nav.eessi.pensjon.gcp.GcpStorageService
-import no.nav.eessi.pensjon.journalforing.JournalforingMedBruker
+import no.nav.eessi.pensjon.journalforing.VurderBrukerInfo
 import no.nav.eessi.pensjon.journalforing.saf.SafClient
 import no.nav.eessi.pensjon.utils.toJson
 import org.junit.jupiter.api.BeforeEach
@@ -71,7 +71,7 @@ internal class SedSendtIntegrationTest : IntegrasjonsBase() {
         fun safClient(): SafClient = SafClient(IntegrasjonsTestConfig().mockedRestTemplate())
 
         @Bean
-        fun journalforingUtenBruker(): JournalforingMedBruker = JournalforingMedBruker(safClient(), gcpStorageService(), mockk(), mockk())
+        fun journalforingUtenBruker(): VurderBrukerInfo = VurderBrukerInfo(safClient(), gcpStorageService(), mockk(), mockk())
         //TODO: Ingen kobling mellomg journalføringservice og denne.. gå gjennom de andre mock -beans også
     }
