+++ src/test/kotlin/no/nav/eessi/pensjon/listeners/SedSendtJournalforingTest.kt
@@ -12,7 +12,7 @@ import no.nav.eessi.pensjon.eux.model.buc.*
 import no.nav.eessi.pensjon.gcp.GcpStorageService
 import no.nav.eessi.pensjon.integrasjonstest.saksflyt.JournalforingTestBase.Companion.FNR_VOKSEN_UNDER_62
 import no.nav.eessi.pensjon.journalforing.JournalforingService
-import no.nav.eessi.pensjon.journalforing.JournalforingMedBruker
+import no.nav.eessi.pensjon.journalforing.VurderBrukerInfo
 import no.nav.eessi.pensjon.journalforing.OpprettJournalPostResponse
 import no.nav.eessi.pensjon.journalforing.OpprettJournalpostRequest
 import no.nav.eessi.pensjon.journalforing.bestemenhet.OppgaveRoutingService
@@ -68,7 +68,7 @@ internal class SedSendtJournalforingTest {
     private val statistikkPublisher = mockk<StatistikkPublisher>(relaxed = true)
     private val navansattKlient = mockk<NavansattKlient>(relaxed = true)
     private val gcpStorageService = mockk<GcpStorageService>()
-    private val journalforingutenbruker = mockk<JournalforingMedBruker>()
+    private val journalforingutenbruker = mockk<VurderBrukerInfo>()
 
     private val jouralforingService = JournalforingService(
         journalpostService,
